Move container functions into a class
